Implement functional shooting; game is now operational.

+++ map.cpp
@@ -13,7 +13,7 @@ const uint16_t MAP_W = 512, MAP_H = 512;
 const uint32_t MAP_A = MAP_W * MAP_H;
                             //eeeeeeee eeeeeeee     llaf ffssssbb
 uint32_t map[MAP_W][MAP_H]; //00000000 00000000 00000000 00000000 - 0000000000000000 entity map id, 0000 RESERVED, 00 luminosity, 0 animated, 000 frame, 0000 sprite, 00 biome
-uint32_t game_time = 0;
+uint64_t game_time = 0;
 
 //Constants
 const uint8_t GEN_ISLANDS = 4;
